backend/requirements.txt
flask 
passporteye 
easyocr 
python-dateutil 
opencv-python 
matplotlib
